Generate integration capability references
Derive provider event and GitHub agent-tool documentation from shipped DTO catalogs so reference pages stay synchronized with runtime capabilities.

Fixes ENG-903
