+++ .changeset/bright-otters-document.md
@@ -1,0 +1,9 @@
+---
+"@shipfox/api-integration-core-dto": minor
+"@shipfox/api-integration-github-dto": minor
+"@shipfox/api-integration-sentry-dto": minor
+"@shipfox/api-integration-webhook-dto": minor
+"@shipfox/api-integration-github": patch
+---
+
+Adds provider event and GitHub agent-tool catalogs for generated integration reference documentation.
