+++ apps/docs/content/docs/integrations/sentry.mdx
@@ -23,19 +23,57 @@ triggers:
 
 ## Events
 
-Shipfox records these Sentry issue events:
+{/* generated:sentry-events:start */}
+### `issue.created`
 
-- `issue.created`
-- `issue.resolved`
-- `issue.assigned`
-- `issue.archived`
-- `issue.unresolved`
+A Sentry issue is created.
 
-Sentry owns the lifecycle meaning and original webhook payload:
+**Emitted when:** Sentry sends an issue webhook with the created action.
 
-<Card title="Sentry issue webhooks" href="https://docs.sentry.io/organization/integrations/integration-platform/webhooks/issues/">
-  The authoritative Sentry issue webhook reference.
-</Card>
+**Payload:** Shipfox-normalized payload.
+
+[Provider payload documentation](https://docs.sentry.io/organization/integrations/integration-platform/webhooks/issues/)
+
+### `issue.resolved`
+
+A Sentry issue is resolved.
+
+**Emitted when:** Sentry sends an issue webhook with the resolved action.
+
+**Payload:** Shipfox-normalized payload.
+
+[Provider payload documentation](https://docs.sentry.io/organization/integrations/integration-platform/webhooks/issues/)
+
+### `issue.assigned`
+
+A Sentry issue is assigned.
+
+**Emitted when:** Sentry sends an issue webhook with the assigned action.
+
+**Payload:** Shipfox-normalized payload.
+
+[Provider payload documentation](https://docs.sentry.io/organization/integrations/integration-platform/webhooks/issues/)
+
+### `issue.archived`
+
+A Sentry issue is archived.
+
+**Emitted when:** Sentry sends an issue webhook with the archived action.
+
+**Payload:** Shipfox-normalized payload.
+
+[Provider payload documentation](https://docs.sentry.io/organization/integrations/integration-platform/webhooks/issues/)
+
+### `issue.unresolved`
+
+A resolved Sentry issue becomes unresolved.
+
+**Emitted when:** Sentry sends an issue webhook with the unresolved action.
+
+**Payload:** Shipfox-normalized payload.
+
+[Provider payload documentation](https://docs.sentry.io/organization/integrations/integration-platform/webhooks/issues/)
+{/* generated:sentry-events:end */}
 
 ## Normalized event payload
