+++ apps/docs/scripts/lib/slug.mjs
@@ -1,0 +1,3 @@
+import GithubSlugger, {slug} from 'github-slugger';
+
+export {GithubSlugger, slug as slugForHeading};
