+++ libs/api/integration/core-dto/src/contracts/index.ts
@@ -1,4 +1,9 @@
 export type {
+  IntegrationEventCatalog,
+  IntegrationEventDoc,
+  IntegrationEventPayloadKind,
+} from './event-catalog.js';
+export type {
   AgentToolCallInput,
   AgentToolCatalogEntry,
   AgentToolCatalogMethod,
