+++ libs/api/integration/github-dto/src/index.ts
@@ -1,1 +1,3 @@
+export * from './agent-tools/catalog.js';
+export * from './event-catalog.js';
 export * from './schemas/index.js';
