+++ libs/api/integration/sentry-dto/src/index.ts
@@ -1,1 +1,2 @@
+export * from './event-catalog.js';
 export * from './schemas/index.js';
