+++ libs/api/integration/webhook-dto/src/constants.ts
@@ -1,0 +1,3 @@
+export const WEBHOOK_PROVIDER = 'webhook' as const;
+export const WEBHOOK_RECEIVED_EVENT = 'received' as const;
+export const WEBHOOK_RESERVED_SLUGS = ['github', 'gitea', 'sentry', 'manual', 'cron'] as const;
